Authorization added for each controller (except sales and invoices due to missing implmententation in this branch), also the error messages are transalted to spanish in order to show those messages on the front

+++ app/Http/Controllers/api/CartController.php
@@ -2,59 +2,114 @@
 
 namespace App\Http\Controllers\api;
 
+use App\Http\Controllers\helpers\RoleAuthorizationHelper;
 use Illuminate\Routing\Controller;
 use Illuminate\Http\Request;
 use App\Models\Product;
 
 class CartController extends Controller
 {
+    protected $roleAuthorizationHelper;
 
     public function __construct()
     {
         $this->middleware('auth:sanctum');
+        $this->roleAuthorizationHelper = new RoleAuthorizationHelper();
     }
 
     public function attachProduct(Request $request)
     {
         $user = $request->user();
+        if (!$this->roleAuthorizationHelper->hasPermission($user->role, 'invoice.update')) {
+            return response()->json([
+                'message' => 'No tienes autorización para realizar esta acción.'
+            ], 401);
+        }
+
         $product = Product::find($request->product_id);
-        error_log($request->product_id);
         $user->products()->attach($product->id, ['quantity' => $request->quantity]);
-        return response()->json(['message' => 'Product attached to user'], 200);
+        return response()->json(['message' => 'Producto agregado al usuario.'], 200);
     }
 
+    /**
+     * Detach a product from the cart.
+     *
+     * @param \Illuminate\Http\Request $request The request instance containing the product details to be detached.
+     * @return \Illuminate\Http\JsonResponse The response indicating the success or failure of the operation.
+     */
     public function detachProduct(Request $request)
     {
         $user = $request->user();
+        if (!$this->roleAuthorizationHelper->hasPermission($user->role, 'invoice.update')) {
+            return response()->json([
+                'message' => 'No tienes autorización para realizar esta acción.'
+            ], 401);
+        }
+
         $product = Product::find($request->product_id);
         $user->products()->detach($product->id);
-        return response()->json(['message' => 'Product detached from user'], 200);
+        return response()->json(['message' => 'Producto eliminado del usuario.'], 200);
     }
 
+    /**
+     * Retrieve the list of products in the cart.
+     *
+     * @param \Illuminate\Http\Request $request The HTTP request instance.
+     * @return \Illuminate\Http\JsonResponse The JSON response containing the list of products.
+     */
     public function getProducts(Request $request)
     {
         $user = $request->user();
+        if (!$this->roleAuthorizationHelper->hasPermission($user->role, 'invoice.read')) {
+            return response()->json([
+                'message' => 'No tienes autorización para realizar esta acción.'
+            ], 401);
+        }
+
         $products = $user->products;
         return response()->json(["data" => $products], 200);
     }
 
+    /**
+     * Update the quantity of a product in the cart.
+     *
+     * @param \Illuminate\Http\Request $request
+     * @return \Illuminate\Http\Response
+     */
     public function updateProductQuantity(Request $request)
     {
         $user = $request->user();
+        if (!$this->roleAuthorizationHelper->hasPermission($user->role, 'invoice.update')) {
+            return response()->json([
+                'message' => 'No tienes autorización para realizar esta acción.'
+            ], 401);
+        }
+
         $product = Product::find($request->product_id);
         $user->products()->updateExistingPivot($product->id, ['quantity' => $request->quantity]);
-        return response()->json(['message' => 'Product quantity updated'], 200);
+        return response()->json(['message' => 'Cantidad del producto actualizada.'], 200);
     }
 
+    /**
+     * Remove all products from the cart.
+     *
+     * @param \Illuminate\Http\Request $request The incoming request instance.
+     * @return \Illuminate\Http\JsonResponse The response indicating the result of the operation.
+     */
     public function cleanProducts(Request $request)
     {
         $user = $request->user();
-        $products = $user->products;
+        if (!$this->roleAuthorizationHelper->hasPermission($user->role, 'invoice.update')) {
+            return response()->json([
+                'message' => 'No tienes autorización para realizar esta acción.'
+            ], 401);
+        }
 
+        $products = $user->products;
         foreach ($products as $product) {
             $user->products()->detach($product->id);
         }
 
-        return response()->json(['message' => 'Products cleaned'], 200);
+        return response()->json(['message' => 'Productos eliminados.'], 200);
     }
 }
